Move webhook move hooks to hooks.go
The conditions hooks are at conditions.go but we are going to have also
some hooks related to node so let's move the current hooks to clear
place.
