+++ pkg/webhook/nodenetworkconfigurationpolicy/conditions.go
@@ -3,8 +3,6 @@ package nodenetworkconfigurationpolicy
 import (
 	corev1 "k8s.io/api/core/v1"
 	logf "sigs.k8s.io/controller-runtime/pkg/runtime/log"
-	"sigs.k8s.io/controller-runtime/pkg/webhook"
-	"sigs.k8s.io/controller-runtime/pkg/webhook/admission"
 
 	shared "github.com/nmstate/kubernetes-nmstate/api/shared"
 	nmstatev1beta1 "github.com/nmstate/kubernetes-nmstate/api/v1beta1"
@@ -31,24 +29,4 @@ func setConditionsUnknown(policy nmstatev1beta1.NodeNetworkConfigurationPolicy) 
 
 func atEmptyConditions(policy nmstatev1beta1.NodeNetworkConfigurationPolicy) bool {
 	return policy.Status.Conditions == nil || len(policy.Status.Conditions) == 0
-}
-
-func deleteConditionsHook() *webhook.Admission {
-	return &webhook.Admission{
-		Handler: admission.HandlerFunc(
-			mutatePolicyHandler(
-				always,
-				deleteConditions,
-			)),
-	}
-}
-
-func setConditionsUnknownHook() *webhook.Admission {
-	return &webhook.Admission{
-		Handler: admission.HandlerFunc(
-			mutatePolicyHandler(
-				atEmptyConditions,
-				setConditionsUnknown,
-			)),
-	}
 }
