+++ pkg/webhook/nodenetworkconfigurationpolicy/timestamp.go
@@ -4,9 +4,6 @@ import (
 	"strconv"
 	"time"
 
-	"sigs.k8s.io/controller-runtime/pkg/webhook"
-	"sigs.k8s.io/controller-runtime/pkg/webhook/admission"
-
 	nmstatev1beta1 "github.com/nmstate/kubernetes-nmstate/api/v1beta1"
 )
 
@@ -21,14 +18,4 @@ func setTimestampAnnotation(policy nmstatev1beta1.NodeNetworkConfigurationPolicy
 	}
 	policy.ObjectMeta.Annotations[TimestampLabelKey] = value
 	return policy
-}
-
-func setTimestampAnnotationHook() *webhook.Admission {
-	return &webhook.Admission{
-		Handler: admission.HandlerFunc(
-			mutatePolicyHandler(
-				always,
-				setTimestampAnnotation,
-			)),
-	}
 }
